+++ src/upgrade/dependency.metadata.ts
